requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
firebase-admin==6.2.0
requests==2.31.0 
openai==0.27.9 
google-api-python-client==2.105.0 
google-auth-httplib2==0.1.1 
google-auth-oauthlib==1.1.0
